+++ src/main/java/com/skypedal/skypedal_backend/rest/LocationController.java
@@ -18,28 +18,28 @@ public class LocationController {
 
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping("")
-    public LocationDTO add(@RequestBody LocationDTO locationDTO, @RequestParam Integer userId) {
+    public LocationDTO add(@RequestBody LocationDTO locationDTO, @RequestParam Long userId) {
         return this.service.add(locationDTO, userId);
     }
 
     @GetMapping("/search")
-    public List<LocationDTO> query(@RequestParam String query, @RequestParam Integer userId) {
+    public List<LocationDTO> query(@RequestParam String query, @RequestParam Long userId) {
         return this.service.query(query, userId);
     }
 
 
     @GetMapping("")
-    public List<LocationDTO> getAll(@RequestParam Integer userId) {
+    public List<LocationDTO> getAll(@RequestParam Long userId) {
         return this.service.get(userId);
     }
 
     @GetMapping("/{id}")
-    public LocationDTO get(@PathVariable Integer id, @RequestParam Integer userId) {
+    public LocationDTO get(@PathVariable Integer id, @RequestParam Long userId) {
         return this.service.getById(id, userId);
     }
 
     @DeleteMapping("/{id}")
-    public LocationDTO delete(@PathVariable Integer id, @RequestParam Integer userId) {
+    public LocationDTO delete(@PathVariable Integer id, @RequestParam Long userId) {
         return this.service.removeById(id, userId);
     }
 }
